Update patch set 2
Patch Set 2: Abandoned

This change on its own doesn't seem to bring any actual functionality, and we have no visibility over a further series of changes that would bring in such functionality.

While having small changes is desirable when doing code reviews, it's still necessary to have some visibility over the end goal and the plan to get there in order to avoid weighing down the code with stubs of features that aren't actually functional.

As such, we're abandoning this change. Thanks for your contributions.

Patch-set: 2