Update patch set 1
Patch Set 1:

Our init.rc does a chown/chmod of rfkill0/rfkill/state so that the bluetooth can tweak it. With this change, the chmod/chown won't work if the enumeration changes.

What is the proper way to make the state file owned by bluetooth so that bluetoothd can open it?

Patch-set: 1