Update patch set 1
Patch Set 1:

Bruce, we (android-x86) had this problem and discussed it deeply before. I think the best solution is to handle the rfkill add uevents. Since vold can receive all uevents, I made a patch to implement that. See
http://git.android-x86.org/?p=platform/system/core.git;a=commitdiff;h=76383ab69ddce989927b2b5651156e2e0f6c0668

But that's for Eclair. We don't have patch for vold2 of Froyo yet.

Patch-set: 1