Update patch set 1
Patch Set 1:

we have few scenarios in ours where we use hciattach in some cases and don't use them in other. (as in the service entry hciattach is run as hciattach -l or something...).
[in the non-HCI-UART case we use TI_ST which creates a hci0 interface on boot like btusb]

So, basically we need to see scenarios like,
1. when hci0 exists during boot, and bluetoothd is run first, HCIDEVUP happens then and there removing the necessity of the ioctl DEVUP from bluedroid.

2. If hci0 is created only after hciattach, then running bluetoothd 1st in fine, but then we need not do an ioctl DEVUP because hci0 does not exist, and if we do, it would throw up and error - since hci0 is only created after running hciattach which would be the next step.

So we wanted to verify these scenarios 1st before submission of that patch.
So, Please review this patch for now and provide comments on the above mentioned scenarios.

Patch-set: 1