Fix the timeout in bt_enable.
1000 * 100ms == 100s, not 10s. This means my AOSP maguro takes 100s
to start system_server after a runtime restart (the initial boot works
fine), which is a very long time to stare at a blank screen. The
intended 10s is perfectly reasonable; we're doing useful work in
parallel for most of that time anyway.

I've also made the logging more like the other logging in this file:

  bluedroid: bt_enable: ioctl(248, HCIDEVUP, HCI_DEV_ID) failed: No such device (19)

(I've no idea why we fail to enable bluetooth after a runtime
restart in AOSP.)

Change-Id: I8321bfed380720e57c43f2f94569ab4f983685f3